% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneExpressionFromGEO.r
\name{dec_two}
\alias{dec_two}
\title{Function that returns numeric values with 2 decimal numbers.}
\usage{
dec_two(x)
}
\arguments{
\item{x}{input numeric value with N decimal numbers.}
}
\value{
a numeric value with 2 decimal numbers.
}
\description{
Function that returns numeric values with 2 decimal numbers.
}
\examples{
aaa <- dec_two(8.31232)
}
