% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneExpressionFromGEO.r
\name{getGeneExpressionFromGEO}
\alias{getGeneExpressionFromGEO}
\title{Function that reads in the GEO code and the platform ID of a dataset, and returns the gene expression dataframe.}
\usage{
getGeneExpressionFromGEO(
  datasetGeoCode,
  datasetPlatform,
  retrieveGeneSymbols,
  verbose = FALSE
)
}
\arguments{
\item{datasetGeoCode}{the GEO code of a dataset.}

\item{datasetPlatform}{the ID of the platform of the dataset.}

\item{retrieveGeneSymbols}{a boolean flag stating if the function should retrieve the gene symbols or not.}

\item{verbose}{a boolean flag stating if helping messages should be printed or not}
}
\value{
a gene expression dataset.
}
\description{
Function that reads in the GEO code and the platform ID of a dataset, and returns the gene expression dataframe.
}
\examples{
geneExpressionDF1 <- getGeneExpressionFromGEO("GSE3268", "GPL96", FALSE, FALSE)
}
